backend-python-memory/requirements.txt
fastapi
uvicorn[standard]
pydantic

chromadb

sentence-transformers

grpcio
grpcio-tools
grpcio-health-checking

opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi

